### 2018-07-01 #799 @danehkar
Based on your list of joss reviewers, I think that @mgalloy @Scivision @Mankoff would be great reviewers for IDL programs.
