### 2018-08-09 #799 @zingale
I'm not really an IDL person anymore, and don't have IDL to test it, so I don't think I can review
