### 2018-08-10 #799 @ziotom78
Same for me, it has been years since I used IDL and no longer have a license.

In data 9 agosto 2018 9:57:12 PM Michael Zingale <notifications@github.com>
ha scritto:
> I'm not really an IDL person anymore, and don't have IDL to test it, so I
> don't think I can review
