### 2018-08-10 #799 @danehkar
It might work on [GNU Data Language (GDL)](https://github.com/gnudatalanguage/gdl). As my library depends on [IDL Astronomy User's library](https://idlastro.gsfc.nasa.gov/homepage.html), I cannot fix any GDL-incompatibility bugs related to IDL Astronomy User's library. I am also preparing a Python version of these package (AtomNeb + proEQUIB). However, I will release them next year. Most people who work at US research institutes have IDL license on their computers. For example, @wlandsman who works on IDL Astronomy User's library,  [IDLAstro](https://github.com/wlandsman/IDLAstro), might be able to test my IDL libraries (AtomNeb and proEQUIB), or @flamingbear who works on [Coyote IDL Library](https://github.com/idl-coyote/coyote).
