### 2018-08-10 #799 @wlandsman
FYI, the team leading the development of GDL (the free IDL clone) have
worked to make it compatible with IDL Astronomy User's library, and have
verified this for the case of the FITS I/O routines.   I took a  quick look
at  AtomNeb and it seems very likely to work under GDL since it does not
use widget or very recent IDL additions.

On Fri, Aug 10, 2018 at 1:12 PM Ashkbiz Danehkar <notifications@github.com>
wrote:

> It might work on GNU Data Language (GDL)
> <https://github.com/gnudatalanguage/gdl>. As my library depends on IDL
> Astronomy User's library <https://idlastro.gsfc.nasa.gov/homepage.html>,
> I cannot fix any GDL-incompatibility bugs related to IDL Astronomy User's
> library. I am also preparing a Python version of these package (AtomNeb +
> proEQUIB). However, I will release them next year. Most people who work at
> US research institutes have IDL license on their computers. For example,
> @wlandsman <https://github.com/wlandsman> who works on IDL Astronomy
> User's library, IDLAstro <https://github.com/wlandsman/IDLAstro>, might
> be able to test my IDL libraries (AtomNeb and proEQUIB), or @flamingbear
> <https://github.com/flamingbear> who works on Coyote IDL Library
> <https://github.com/idl-coyote/coyote>.
>
