### 2018-08-16 #799 @arfon
Hi @mgalloy & @wlandsman, firstly, many thanks for your help/interest in these packages thus far.

We've recently had two submissions from @danehkar: this one, and the [proEQUIB](https://github.com/openjournals/joss-reviews/issues/798) IDL library.

I'm wondering if you'd be both willing to review _both_ of these packages? I ask as we're really struggling to find suitable reviewers for these packages.

If this is something that you'd be open to then I would go ahead and start the reviews for both but there would be no expectation on our part that you complete these both simultaneously!

What do you both think?
