### 2018-08-17 #898 @whedon
[REVIEW]: AtomNeb: IDL Library for Atomic Data of Ionized Nebulae
**Submitting author:** @danehkar (<a href="http://orcid.org/0000-0003-4552-5997">Ashkbiz Danehkar</a>)
**Repository:** <a href="https://github.com/atomneb/AtomNeb-idl" target ="_blank">https://github.com/atomneb/AtomNeb-idl</a>
**Version:** v0.0.2
