### 2018-08-20 atomneb/AtomNeb-idl#1 @mdpiper
Add information for contributors
Information should be provided to direct users how to contribute, report problems, or seek help when working with this software.

For example, [astropy](https://github.com/astropy/astropy) has a [CONTRIBUTING.md](https://github.com/astropy/astropy/blob/master/CONTRIBUTING.md) file that provides this information. A (far shorter!) version of such a file be appropriate here.
