### 2018-08-20 #898 @mdpiper
Info for contributors (last item in the **Documentation** section) is requested in https://github.com/atomneb/AtomNeb-idl/issues/1.
