### 2018-08-20 atomneb/AtomNeb-idl#2 @mdpiper
Add unit tests
Unit tests, written by the developer, are extremely helpful for ensuring that software behaves as a developer intends. (I've caught *so* many defects through unit tests.) Unit tests should be added to this package.
