### 2018-08-20 atomneb/AtomNeb-idl#2 @mdpiper
A suggestion: [mgunit](https://github.com/mgalloy/mgunit) is a mature and widely used unit testing framework for IDL.
