### 2018-08-24 atomneb/AtomNeb-idl#3 @mdpiper
Create a public release
The repo should be stamped with a version number (e.g., in [AtomNeb-idl/releases](https://github.com/atomneb/AtomNeb-idl/releases)) for public release.
