### 2018-08-24 #898 @mdpiper
The package isn't stamped with a release, so I created https://github.com/atomneb/AtomNeb-idl/issues/3 to request one. This addresses the third item in the **General checks** checklist.
