### 2018-08-24 atomneb/AtomNeb-idl#4 @mdpiper
Create API documentation
Although each function in AtomNeb-idl has a docstring describing its functionality, it's a best practice to create separate API documentation for a package. This makes it easier for developers (and users) to understand what each routine in the package does.

A recommendation: [IDLdoc](https://github.com/mgalloy/idldoc) produces nice API documentation, and it's an accepted standard in the IDL community.
