### 2018-08-24 #898 @mdpiper
The package doesn't include API documentation (the fourth item in the **Documentation** checklist). I've requested this in https://github.com/atomneb/AtomNeb-idl/issues/4.
