### 2018-08-30 #898 @mdpiper
I offered a PR (https://github.com/atomneb/AtomNeb-idl/pull/5) to automatically locate the package data from the examples.
