### 2018-08-30 atomneb/AtomNeb-idl#6 @mdpiper
Add note on installing package
IDL's path management can be difficult, especially for a newcomer. It would be helpful to add a short note to the README about how to install this package. For example:

### Installation

To install the **AtomNeb** database and API library on your machine, simply add this directory to your IDL path. If you're new to IDL, learn more about path management in [this article](https://www.harrisgeospatial.com/Support/SelfHelpTools/HelpArticles/HelpArticles-Detail/TabId/2718/ArtMID/10220/ArticleID/16156/Quick-tips-for-customizing-your-IDL-program-search-path.aspx) from Harris Geospatial Solutions, or in [this article](http://www.idlcoyote.com/code_tips/installcoyote.php) by David Fanning of the Coyote IDL Library.
