### 2018-08-30 #898 @mdpiper
Created an issue (https://github.com/atomneb/AtomNeb-idl/issues/6) to address package installation (the first item in the **Functionality** checklist).
