### 2018-09-06 #898 @mdpiper
My review of **AtomNeb** is complete. The library and its associated paper are appropriate for publication in JOSS, pending  @danehkar's responses to the minor issues I've opened in the repository.

(@arfon I'll complete the open items in the review checklist once these issues have been addressed.)
