### 2018-09-11 atomneb/AtomNeb-idl#4 @mgalloy
If you do this for the code in your `examples` directory as well, you can satisfy the documentation of real-world examples requirement as well.
