### 2018-09-11 #898 @mgalloy
Catching up after some travel. Added some comments on issues opened by @mdpiper.

I think it would also be useful to add a short sentence after the opening description of AtomNeb database in your [README](https://github.com/atomneb/AtomNeb-idl/blob/master/README.md) that indicates precisely how `AtomNeb-idl` relates to it. Does it contain the data, access them via the web, access data you have generated, etc.?

I should be able to start looking at proEQUIB shortly.
