### 2018-09-11 atomneb/AtomNeb-idl#6 @mgalloy
Also, are there requirements for installation? a minimum version of IDL?
