### 2018-11-24 f3ce3fb
Note on Installation and minimum IDL version
Add note on Installation and minimum IDL version suggested by @mdpiper and @mgalloy
