### 2018-11-24 atomneb/AtomNeb-idl#6 @danehkar
I added a [note](https://github.com/atomneb/AtomNeb-idl/commit/f3ce3fb53fb3d741a7fb17e8ce9a7c35f11fddd1) to README on Installation and minimum IDL version.
