### 2018-11-25 930f192
CONTRIBUTING.md added
CONTRIBUTING.md, which contains guidelines for contributors, are included in this IDL library @mdpiper
