### 2018-11-25 atomneb/AtomNeb-idl#1 @danehkar
I included [CONTRIBUTING.md](https://github.com/atomneb/AtomNeb-idl/blob/master/CONTRIBUTING.md) in the [AtomNeb-idl](https://github.com/atomneb/AtomNeb-idl) repository that contains some guidelines for contributors.
