### 2018-11-27 #898 @danehkar
> @mdpiper Info for contributors (the last item in the **Documentation** checklist) isn't present, so I made a request in [atomneb/AtomNeb-idl#1](https://github.com/atomneb/AtomNeb-idl/issues/1).

I included [CONTRIBUTING.md](https://github.com/atomneb/AtomNeb-idl/blob/master/CONTRIBUTING.md) in the [AtomNeb-idl](https://github.com/atomneb/AtomNeb-idl) repository that contains some guidelines for contributors. Now I closed the issue [atomneb/AtomNeb-idl#1](https://github.com/atomneb/AtomNeb-idl/issues/1).
