### 2018-11-27 #898 @danehkar
> @mdpiper I offered a PR ([atomneb/AtomNeb-idl#5](https://github.com/atomneb/AtomNeb-idl/pull/5)) to automatically locate the package data from the examples.

Thank you! I merged your PR into the repository. Now the package data are automatically located in the examples: [atomneb/AtomNeb-idl/examples](https://github.com/atomneb/AtomNeb-idl/tree/master/examples).
