### 2018-11-27 #898 @danehkar
> @mdpiper Created an issue ([atomneb/AtomNeb-idl#6](https://github.com/atomneb/AtomNeb-idl/issues/6)) to address package installation (the first item in the **Functionality** checklist).

Thank you for your installation note suggestion. I added an installation note in [README.md](https://github.com/atomneb/AtomNeb-idl/blob/master/README.md):

> To install the **AtomNeb** database and its API library, simply add the path of this package directory to your IDL path. This package requires IDL version 7.1 or later. For more information about the path management in IDL, read [the tips for customizing IDL program path](https://www.harrisgeospatial.com/Support/Self-Help-Tools/Help-Articles/Help-Articles-Detail/ArtMID/10220/ArticleID/16156/Quick-tips-for-customizing-your-IDL-program-search-path) provided by Harris Geospatial Solutions or [the IDL library installation note](http://www.idlcoyote.com/code_tips/installcoyote.php) by David Fanning in the Coyote IDL Library.

Now I closed the issue [atomneb/AtomNeb-idl#6](https://github.com/atomneb/AtomNeb-idl/issues/6).
