### 2018-11-28 #898 @mdpiper
@danehkar With your recent updates, I've marked as complete a few more items on the checklist.
