### 2019-01-25 #898 @danehkar
@labarba I am currently working on those issues. Hopefully, I will resolve them in a month. Thank you for your patience.
