### 2019-02-19 c8e99e4
add appveyor + travis
