+++ README.rst
@@ -1,0 +1,132 @@
+=======
+AtomNeb (IDL version)
+=======
+    
+.. image:: https://travis-ci.org/equib/AtomNeb-idl.svg?branch=master
+    :target: https://travis-ci.org/atomneb/AtomNeb-idl
+    :alt: Build Status
+
+.. image:: https://ci.appveyor.com/api/projects/status/at7fd0xbr1e9u75g?svg=true
+    :target: https://ci.appveyor.com/project/danehkar/atomneb-idl
+    :alt: Build Status
+
+.. image:: http://mybinder.org/badge.svg
+    :target: http://mybinder.org/repo/atomneb/atomneb-idl
+    :alt: Binder
+
+.. image:: https://img.shields.io/aur/license/yaourt.svg
+    :target: https://github.com/atomneb/AtomNeb-idl/blob/master/LICENSE
+    :alt: GitHub license
+
+Description
+============
+
+Collisionally Excited Lines
+---------------------------
+
+The **AtomNeb** database for *collisionally excited lines*  contains sets of `atomic datasets <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data>`_, which were generated for use in `proEQUIB <https://github.com/equib/proEQUIB>`_, `pyEQUIB <https://github.com/equib/pyEQUIB>`_, and other nebular spectral analysis tools. They include energy levels (Ej), collision strengths (Ωij), and transition probabilities (Aij) of the most ions commonly observed in ionized nebulae.
+
+The atomic datasets for collisionally excited lines are as follows:
+
+* `Collection <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/collection>`_ from the `National Institute of Standards and Technology (NIST) Atomic Spectra Database <https://www.nist.gov/pml/atomic-spectra-database>`_, the `CHIANTI atomic database <http://www.chiantidatabase.org/>`_, and some improved atomic data from Cloudy v13.04 and pyNeb v1.0.
+
+* `Chianti52 <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti52>`_ from the `CHIANTI atomic database <http://www.chiantidatabase.org/>`_ version 5.2.
+
+* `Chianti60 <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti60>`_ from the `CHIANTI atomic database <http://www.chiantidatabase.org/>`_ version 6.0.
+
+* `Chianti70 <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti70>`_ from the `CHIANTI atomic database <http://www.chiantidatabase.org/>`_ version 7.0.
+
+
+Each dataset contains the following `atomic data FITS files <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti70>`_:
+
+* **AtomElj.fits** contains *Energy Levels* (Ej).
+
+* **AtomOmij.fits** contains *Collision Strengths* (Ωij).
+
+* **AtomAij.fits** contains *Transition Probabilities* (Aij).
+
+
+Recombination Lines
+-------------------
+
[…]
