+++ .appveyor.yml
@@ -1,0 +1,29 @@
+image:
+#- Visual Studio 2017
+- Ubuntu
+
+clone_depth: 3
+
+build: off
+
+init:
+- sh: sudo apt-get -yq update > /dev/null
+- sh: sudo apt-get install -yq --no-install-suggests --no-install-recommends gnudatalanguage > /dev/null
+
+install:
+  - export DISPLAY=':99.0'
+  - Xvfb :99 -screen 0 1024x768x24 > /dev/null 2>&1 &
+  - git submodule update --init --recursive
+  - export GDL_STARTUP=/home/appveyor/projects/atomneb-idl/test_appveyor/.gdl_startup
+
+test_script:
+- gdl -e @test_appveyor/atomneb_chianti52_example.pro
+- gdl -e @test_appveyor/atomneb_chianti60_example.pro
+- gdl -e @test_appveyor/atomneb_chianti70_example.pro
+- gdl -e @test_appveyor/atomneb_collection_example.pro
+- gdl -e @test_appveyor/atomneb_rc_ppb91_example.pro
+- gdl -e @test_appveyor/atomneb_rc_sh95_example.pro
+- gdl -e @test_appveyor/atomneb_rc_collection_example.pro
+- gdl -e @test_appveyor/atomneb_rc_he_i_pfsd12_example.pro
+- gdl -e @test_appveyor/atomneb_rc_n_ii_fsl13_example.pro
+
