+++ README.md
@@ -1,53 +1,0 @@
-## AtomNeb (IDL version)
-[![GitHub license](https://img.shields.io/aur/license/yaourt.svg)](https://github.com/atomneb/AtomNeb-idl/blob/master/LICENSE)
-
-**Atomic Data for Ionized Nebulae**
-
-
-### Description
-
-The **AtomNeb** database for ***collisionally excited lines*** contains sets of [atomic datasets](https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data), which were generated for use in [proEQUIB](https://github.com/equib/proEQUIB), [pyEQUIB](https://github.com/equib/pyEQUIB), and other nebular spectral analysis tools. They include energy levels (Ej), collision strengths (Ωij), and transition probabilities (Aij) of the most ions commonly observed in ionized nebulae.
-
-The atomic datasets for collisionally excited lines are as follows:
-
-* **[Collection](https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/collection)** from the [National Institute of Standards and Technology (NIST) Atomic Spectra Database](https://www.nist.gov/pml/atomic-spectra-database), the [CHIANTI atomic database](http://www.chiantidatabase.org/), and some improved atomic data from Cloudy v13.04 and pyNeb v1.0.
-
-* **[Chianti52](https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti52)** from the [CHIANTI atomic database](http://www.chiantidatabase.org/) version 5.2.
-
-* **[Chianti60](https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti60)** from the [CHIANTI atomic database](http://www.chiantidatabase.org/) version 6.0.
-
-* **[Chianti70](https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti70)** from the [CHIANTI atomic database](http://www.chiantidatabase.org/) version 7.0.
-
-Each dataset contains the following [atomic data FITS files](https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti70):
-
-* **AtomElj.fits** contains *Energy Levels* (Ej).
-
-* **AtomOmij.fits** contains *Collision Strengths* (Ωij).
-
-* **AtomAij.fits** contains *Transition Probabilities* (Aij).
-
-The **AtomNeb** database for ***recombination lines*** contains sets of [recombination coefficients](https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data-rc), which were generated for use in [proEQUIB](https://github.com/equib/proEQUIB), [pyEQUIB](https://github.com/equib/pyEQUIB), and other nebular spectral analysis tools. They include effective recombination coefficient (αeff) of recombination lines of H I, He I, He II, C I, C II, C III, C VI, N II, N III, N IV, N V, N VI, N VII, O II, O III, O IV, O V, O VI, O VIII, and Ne II ions typically observed in ionized nebulae, as well as Branching ratios (Br) of O II and N II lines.
-
-The atomic datasets for recombination lines are as follows:
-
[…]
