+++ README.rst
@@ -1,5 +1,5 @@
 =======
-AtomNeb (IDL version)
+AtomNeb (IDL/GDL)
 =======
     
 .. image:: https://travis-ci.org/atomneb/AtomNeb-idl.svg?branch=master
@@ -21,6 +21,9 @@ AtomNeb (IDL version)
 Description
 ============
 
+**AtomNeb-idl** is an `IDL <http://www.harrisgeospatial.com/ProductsandSolutions/GeospatialProducts/IDL.aspx>`_/`GDL <http://gnudatalanguage.sourceforge.net/>`_ library for reading atomic data from the `AtomNeb` database for *collisionally excited lines* and *recombination lines* typically observed in spectra of ionized gaseous nebulae.
+
+
 Collisionally Excited Lines
 ---------------------------
 
@@ -28,13 +31,13 @@ The **AtomNeb** database for *collisionally excited lines*  contains sets of `at
 
 The atomic datasets for collisionally excited lines are as follows:
 
-* `Collection <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/collection>`_ from the `National Institute of Standards and Technology (NIST) Atomic Spectra Database <https://www.nist.gov/pml/atomic-spectra-database>`_, the `CHIANTI atomic database <http://www.chiantidatabase.org/>`_, and some improved atomic data from Cloudy v13.04 and pyNeb v1.0.
+* `Collection <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/collection>`_ from the `National Institute of Standards and Technology (NIST) Atomic Spectra Database <https://www.nist.gov/pml/atomic-spectra-database>`_, the `CHIANTI atomic database <http://www.chiantidatabase.org/>`_, and some improved atomic data from `Cloudy v13.04 <https://www.nublado.org/>`_ and pyNeb v1.0. This collection was compiled according to the atomic data used in `pyNeb v1.0 <http://www.iac.es/proyecto/PyNeb/>`_.
 
-* `Chianti52 <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti52>`_ from the `CHIANTI atomic database <http://www.chiantidatabase.org/>`_ version 5.2.
+* `Chianti52 <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti52>`_ from the `CHIANTI atomic database <http://www.chiantidatabase.org/>`_ version 5.2. This dataset was compiled according to the atomic data used in `MOCASSIN <https://github.com/mocassin/MOCASSIN-2.0>`_.
 
-* `Chianti60 <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti60>`_ from the `CHIANTI atomic database <http://www.chiantidatabase.org/>`_ version 6.0.
+* `Chianti60 <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti60>`_ from the `CHIANTI atomic database <http://www.chiantidatabase.org/>`_ version 6.0. This dataset was compiled according to the atomic data used in `MOCASSIN <https://github.com/mocassin/MOCASSIN-2.0>`_.
 
-* `Chianti70 <https://github.com/atomneb/AtomNeb-idl/tree/master/atomic-data/chianti70>`_ from the `CHIANTI atomic database <http://www.chiantidatabase.org/>`_ version 7.0.
[…]
