+++ docs/README.md
@@ -1,0 +1,4 @@
+## Documentation
+
+The [API Documentation](https://atomneb.github.io/AtomNeb-idl/doc/) is available on [atomneb.github.io/AtomNeb-idl](https://atomneb.github.io/AtomNeb-idl/). The documentation for AtomNeb is produced by ``atomneb_build_doc.pro`` which uses [IDLdoc](https://github.com/mgalloy/idldoc). 
+
