### 2019-02-19 atomneb/AtomNeb-idl#4 @danehkar
I created API documentation with [IDLdoc](https://github.com/mgalloy/idldoc) and uploaded them to [docs](https://atomneb.github.io/AtomNeb-idl/doc/).
