### 2019-02-19 #898 @danehkar
> The package doesn't include API documentation (the fourth item in the **Documentation** checklist). I've requested this in [atomneb/AtomNeb-idl#4](https://github.com/atomneb/AtomNeb-idl/issues/4).

@mdpiper The documentation is now hosted at https://atomneb.github.io/AtomNeb-idl/doc/
