### 2019-02-20 atomneb/AtomNeb-idl#2 @danehkar
@mdpiper I added  [mgunit](https://github.com/mgalloy/mgunit) in the folder  [test_mgunit](https://github.com/atomneb/AtomNeb-idl/tree/master/test_mgunit) for automated tests.

To test, you can download AtomNeb-idl with its dependencies as follows

    git clone --recursive https://github.com/atomneb/AtomNeb-idl.git

You need to unpack rc_o_iii_SSB17.fits.tar.gz in ``atomic-data-rc``:

    tar -xvf rc_o_iii_SSB17.fits.tar.gz

To run the test, you need to run the following command:

    idl test_all.pro

There should not be any failure in the ``test-results.html`` file.
