### 2019-02-20 #898 @mgalloy
I ran the tests and confirm they all passed. One note: I had to go get `mgutlibtestcase__define.pro` — it is not part of `mgunit`. I would add it to the `test_mgunit` directory, possibly renaming to `atomnebuttestcase__define.pro`.
