### 2019-02-21 #898 @danehkar
> I ran the tests and confirm they all passed. One note: I had to go get `mgutlibtestcase__define.pro` — it is not part of `mgunit`. I would add it to the `test_mgunit` directory, possibly renaming to `atomnebuttestcase__define.pro`.

Thank you for testing. I will copy `mgutlibtestcase__define.pro`to the `test_mgunit` directory, and rename it  to `atomnebuttestcase__define.pro`.
