### 2019-02-21 #898 @danehkar
@arfon I would like to rename the _title_ of this submission ([898](https://github.com/openjournals/joss-reviews/issues/898)) from **AtomNeb: Atomic Data for Ionized Nebulae** to **AtomNeb: IDL Library for Atomic Data of Ionized Nebulae** since I am working on a similar python package, and I want to distinguish between them. I will upload the revised paper of this IDL library soon.
