+++ paper/paper.md
@@ -1,39 +1,42 @@
 ---
-title: 'AtomNeb: Atomic Data for Ionized Nebulae'
+title: 'AtomNeb: IDL Library for Atomic Data of Ionized Nebulae'
 tags:
-  - IDL
-  - atomic data
-  - spectral analysis
-  - ionized nebulae
   - astrophysics
+  - atomic data
+  - gaseous nebulae
+  - spectral analysis
+  - IDL
+  - GDL
 authors:
   - name: Ashkbiz Danehkar
     orcid: 0000-0003-4552-5997
-    affiliation: "1, 2"
+    affiliation: "1, 2, 3"
 affiliations:
  - name: Department of Physics and Astronomy, Macquarie University, Sydney, NSW 2109, Australia
    index: 1
  - name: Harvard-Smithsonian Center for Astrophysics, 60 Garden Street, Cambridge, MA 02138, USA 
    index: 2
-date: 29 June 2018
+ - name: Department of Astronomy, University of Michigan, Ann Arbor, MI 48109, USA 
+   index: 3
+date: 30 June 2018
 bibliography: paper.bib
 ---
 
 # Summary
 
-Ionized nebulae are used as an astrophysical tool to trace of the chemical 
+An ionized gaseous nebula can be used as an astrophysical tool to trace of the chemical 
 composition of the interstellar medium in the Milky Way and other galaxies, 
-and understand stellar mixing processes. Temperatures, densities, and ionic 
-abundances are determined from _collisionally excited lines_ by solving statistical 
+and to understand stellar mixing processes. Temperatures, densities, and ionic 
+abundances can be determined from _collisionally excited lines_ (CEL) by solving statistical 
 equilibrium equations using collision strengths and transition probabilities of ions. 
-Similarly, physical conditions and ionic abundances are calculated from 
-_recombination lines_ using recombination coefficients of ions. Hence, the 
-atomic data for both the collisional excitation and recombination process are 
-important to determine physical conditions and elemental abundances of ionized 
+Moreover, physical conditions and ionic abundances can be calculated from 
+_recombination lines_ (RL) using recombination coefficients of ions. Hence, the 
+atomic data for collisional excitation and recombination process are 
+necessary to determine physical conditions and elemental abundances of ionized 
 nebulae from collisionally excited lines and recombination 
-lines [see e.g. @Danehkar:2013; @Danehkar:2014; @Danehkar:2014b; @Danehkar:2016; @Danehkar:2018].
-Moreover, atomic data are essential for calculations of level populations and line emissivities 
-in photoionization codes [e.g. @Ferland:1998; @Kallman:2001; @Ercolano:2003; @Ercolano:2005]. 
+lines (see e.g. @Danehkar:2013; @Danehkar:2014; @Danehkar:2014b; @Danehkar:2016; @Danehkar:2018a).
+Additionally, atomic data are essential for calculations of level populations and line emissivities 
+in photoionization codes (e.g. @Ferland:1998; @Kallman:2001; @Ercolano:2003; @Ercolano:2005). 
 
 ``AtomNeb`` is a database containing atomic data stored in the Flexible Image 
 Transport System (FITS) file format produced for nebular spectral analysis, 
@@ -41,32 +44,41 @@ plasma diagnostics, and chemical abundance analysis. FITS tables
 provide easy access to atomic data for spectral analysis tools. Specifically, 
 ``AtomNeb`` includes the atomic data for both the _collisional excitation_ and 
 _recombination_ process of ions typically observed in nebular astrophysics. 
-It is equipped with application programming interface (API) functions 
-in the Interactive Data Language (IDL) to read the atomic data stored in FITS files.
[…]
