+++ paper/paper.bib
@@ -2,7 +2,7 @@
    author = {{Landi}, E. and {Del Zanna}, G. and {Young}, P.~R. and {Dere}, K.~P. and 
 	{Mason}, H.~E. and {Landini}, M.},
     title = "{CHIANTI-An Atomic Database for Emission Lines. VII. New Data for X-Rays and Other Improvements}",
-  journal = {\apjs},
+  journal = {The Astrophysical Journal Supplement Series},
  keywords = {Atomic Data, Plasmas},
      year = 2006,
     month = jan,
@@ -17,7 +17,7 @@
    author = {{Dere}, K.~P. and {Landi}, E. and {Young}, P.~R. and {Del Zanna}, G. and 
 	{Landini}, M. and {Mason}, H.~E.},
     title = "{CHIANTI - an atomic database for emission lines. IX. Ionization rates, recombination rates, ionization equilibria for the elements hydrogen through zinc and updated atomic data}",
-  journal = {\aap},
+  journal = {Astronomy and Astrophysics},
  keywords = {atomic data, atomic processes, radiation mechanisms: thermal, plasmas},
      year = 2009,
     month = may,
@@ -32,7 +32,7 @@
    author = {{Landi}, E. and {Del Zanna}, G. and {Young}, P.~R. and {Dere}, K.~P. and 
 	{Mason}, H.~E.},
     title = "{CHIANTI{\mdash}An Atomic Database for Emission Lines. XII. Version 7 of the Database}",
-  journal = {\apj},
+  journal = {The Astrophysical Journal},
  keywords = {atomic data, atomic processes, Sun: UV radiation, Sun: X-rays, gamma rays, ultraviolet: general, X-rays: general},
      year = 2012,
     month = jan,
@@ -47,7 +47,7 @@
 @article{Luridiana:2015,
    author = {{Luridiana}, V. and {Morisset}, C. and {Shaw}, R.~A.},
     title = "{PyNeb: a new tool for analyzing emission lines. I. Code description and validation of results}",
-  journal = {\aap},
+  journal = {Astronomy and Astrophysics},
 archivePrefix = "arXiv",
    eprint = {1410.6662},
  primaryClass = "astro-ph.IM",
@@ -65,7 +65,7 @@ archivePrefix = "arXiv",
 @article{Davey:2000,
    author = {{Davey}, A.~R. and {Storey}, P.~J. and {Kisielius}, R.},
     title = "{Recombination coefficients for C II lines}",
-  journal = {\aaps},
+  journal = {Astronomy and Astrophysicss},
  keywords = {ATOMIC DATA, LINE: FORMATION, $\backslash$ION$\{$H$\}$$\{$II$\}$ REGIONS, PLANETARY NEBULAE: GENERAL},
      year = 2000,
     month = feb,
@@ -79,7 +79,7 @@ archivePrefix = "arXiv",
 @article{Escalante:1990,
    author = {{Escalante}, V. and {Victor}, G.~A.},
     title = "{Effective recombination coefficients of neutral carbon and singly ionized nitrogen}",
-  journal = {\apjs},
+  journal = {The Astrophysical Journal Supplement Series},
  keywords = {Atomic Recombination, Carbon, Ionized Gases, Nitrogen Ions, Recombination Coefficient, Forbidden Bands, Heavy Elements, Line Spectra, Radiative Recombination},
      year = 1990,
     month = jul,
@@ -93,7 +93,7 @@ archivePrefix = "arXiv",
 @article{Storey:1994,
    author = {{Storey}, P.~J.},
     title = "{Recombination coefficients for O II lines at nebular temperatures and densities}",
[…]

[…]
