### 2019-02-26 #898 @danehkar
> I ran the tests and confirm they all passed. One note: I had to go get `mgutlibtestcase__define.pro` — it is not part of `mgunit`. I would add it to the `test_mgunit` directory, possibly renaming to `atomnebuttestcase__define.pro`.

@mgalloy Thank you for the correction. I added it to `test_mgunit` directory.
