+++ paper/paper.md
@@ -24,46 +24,48 @@ bibliography: paper.bib
 
 # Summary
 
-An ionized gaseous nebula can be used as an astrophysical tool to trace of the chemical 
-composition of the interstellar medium in the Milky Way and other galaxies, 
-and to understand stellar mixing processes. Temperatures, densities, and ionic 
-abundances can be determined from _collisionally excited lines_ (CEL) by solving statistical 
-equilibrium equations using collision strengths and transition probabilities of ions. 
-Moreover, physical conditions and ionic abundances can be calculated from 
-_recombination lines_ (RL) using recombination coefficients of ions. Hence, the 
-atomic data for collisional excitation and recombination process are 
-necessary to determine physical conditions and elemental abundances of ionized 
-nebulae from collisionally excited lines and recombination 
+An ionized gaseous nebula can be used as an astrophysical tool to trace the chemical 
+composition of the interstellar medium in our Galaxy and other galaxies, 
+and to study mixing processes in stellar evolution. Electron temperatures, electron densities, 
+and ionic abundances can be determined from _collisionally excited lines_ (CEL) by solving statistical 
+equilibrium equations using collision strengths ($\Omega_{ij}$) and transition probabilities ($A_{ij}$) of ions. 
+Moreover, physical conditions and chemical abundances can be calculated from 
+_recombination lines_ (RL) using effective recombination coefficients ($\alpha_{\rm eff}$) of ions. 
+The atomic data (e.g. $\Omega_{ij}$, $A_{ij}$, and $\alpha_{\rm eff}$) 
+are used to calculate line emissivities in nebular spectral analysis tools 
+(e.g. @Howarth:1981; @Shaw:1994; @Shaw:1998; @Luridiana:2015; @Howarth:2016; @Danehkar:2018b), 
+and photoionization codes (e.g. @Ferland:1998; @Kallman:2001; @Ercolano:2003; @Ercolano:2005; @Ercolano:2008). 
+Hence, the atomic data for collisional excitation and recombination process are 
+essential to determine physical conditions and elemental abundances of ionized 
+nebulae from collisionally excited and recombination 
 lines (see e.g. @Danehkar:2013; @Danehkar:2014; @Danehkar:2014b; @Danehkar:2016; @Danehkar:2018a).
-Additionally, atomic data are essential for calculations of level populations and line emissivities 
-in photoionization codes (e.g. @Ferland:1998; @Kallman:2001; @Ercolano:2003; @Ercolano:2005). 
 
 ``AtomNeb`` is a database containing atomic data stored in the Flexible Image 
-Transport System (FITS) file format produced for nebular spectral analysis, 
-plasma diagnostics, and chemical abundance analysis. FITS tables 
-provide easy access to atomic data for spectral analysis tools. Specifically, 
+Transport System (FITS; @Wells:1981; @Hanisch:2001; @Pence:2010) file format 
+produced for nebular spectral analysis. 
+FITS tables provide easy access to atomic data for spectral analysis tools. Especially, 
 ``AtomNeb`` includes the atomic data for both the _collisional excitation_ and 
 _recombination_ process of ions typically observed in nebular astrophysics. 
-The ``AtomNeb-idl`` IDL library is equipped with several application programming 
+The ``AtomNeb`` interface library is equipped with several application programming 
 interface (API) functions written in the Interactive Data Language (IDL) 
-for reading the atomic data from ``AtomNeb`` FITS files. This IDL library 
-can also be employed in the GNU Data Language (GDL) (@Arabas:2010; @Coulais:2010), 
[…]
