+++ paper/paper.bib
@@ -5,7 +5,6 @@
   journal = {The Astrophysical Journal Supplement Series},
  keywords = {Atomic Data, Plasmas},
      year = 2006,
-    month = jan,
    volume = 162,
     pages = {261-280},
       doi = {10.1086/498148},
@@ -20,7 +19,6 @@
   journal = {Astronomy and Astrophysics},
  keywords = {atomic data, atomic processes, radiation mechanisms: thermal, plasmas},
      year = 2009,
-    month = may,
    volume = 498,
     pages = {915-929},
       doi = {10.1051/0004-6361/200911712},
@@ -35,7 +33,6 @@
   journal = {The Astrophysical Journal},
  keywords = {atomic data, atomic processes, Sun: UV radiation, Sun: X-rays, gamma rays, ultraviolet: general, X-rays: general},
      year = 2012,
-    month = jan,
    volume = 744,
       eid = {99},
     pages = {99},
@@ -53,7 +50,6 @@ archivePrefix = "arXiv",
  primaryClass = "astro-ph.IM",
  keywords = {methods: numerical, atomic data, Hii regions, planetary nebulae: general, ISM: abundances},
      year = 2015,
-    month = jan,
    volume = 573,
       eid = {A42},
     pages = {A42},
@@ -68,7 +64,6 @@ archivePrefix = "arXiv",
   journal = {Astronomy and Astrophysicss},
  keywords = {ATOMIC DATA, LINE: FORMATION, $\backslash$ION$\{$H$\}$$\{$II$\}$ REGIONS, PLANETARY NEBULAE: GENERAL},
      year = 2000,
-    month = feb,
    volume = 142,
     pages = {85-94},
       doi = {10.1051/aas:2000139},
@@ -82,7 +77,6 @@ archivePrefix = "arXiv",
   journal = {The Astrophysical Journal Supplement Series},
  keywords = {Atomic Recombination, Carbon, Ionized Gases, Nitrogen Ions, Recombination Coefficient, Forbidden Bands, Heavy Elements, Line Spectra, Radiative Recombination},
      year = 1990,
-    month = jul,
    volume = 73,
     pages = {513-553},
       doi = {10.1086/191479},
@@ -96,7 +90,6 @@ archivePrefix = "arXiv",
   journal = {Astronomy and Astrophysics},
  keywords = {Atomic Spectra, Coefficients, Electron Recombination, H Ii Regions, Line Spectra, Oxygen Spectra, Photoionization, Planetary Nebulae, Radiative Recombination, Configuration Interaction, Electron Density (Concentration), Electron Energy, Electron Orbitals},
      year = 1994,
-    month = feb,
    volume = 282,
     pages = {999-1013},
    adsurl = {http://adsabs.harvard.edu/abs/1994A%26A...282..999S},
@@ -110,7 +103,6 @@ archivePrefix = "arXiv",
   journal = {Monthly Notices of the Royal Astronomical Society},
  keywords = {ATOMIC DATA-ATOMIC PROCESSES-ISM, ABUNDANCES-PLANETARY NUBULAE, INDIVIDUAL, NGC 7009.},
      year = 1995,
-    month = jan,
    volume = 272,
     pages = {369-388},
       doi = {10.1093/mnras/272.2.369},
@@ -125,7 +117,6 @@ archivePrefix = "arXiv",
   journal = {Astronomy and Astrophysicss},
  keywords = {ATOMIC DATA, H II REGIONS, PLANETARY NEBULAE: GENERAL},
      year = 1998,
-    month = dec,
    volume = 133,
     pages = {257-269},
[…]

[…]
