### 2019-02-27 #898 @danehkar
@arfon I updated the paper, and regenerated the pdf.  You can check it here: [799](https://github.com/openjournals/joss-reviews/issues/799).
