### 2019-03-01 #898 @danehkar
@mgalloy Thank you very much for your test and review of this package. I look forward to @mdpiper's comments.

@arfon Thank you for regenerating the article proof. It looks fine.
