### 2019-03-01 #898 @mdpiper
@danehkar With your recent updates, I've marked off the last three items on my checklist, completing my review. One minor issue, which I defer to @arfon, is the version on the software (v0.0.1) doesn't quite follow semantic versioning. If this isn't an issue for JOSS, it isn't for me either.
