### 2019-03-02 #898 @arfon
> @danehkar With your recent updates, I've marked off the last three items on my checklist, completing my review. One minor issue, which I defer to @arfon, is the version on the software (v0.0.1) doesn't quite follow semantic versioning. If this isn't an issue for JOSS, it isn't for me either.

We can update the version at the end of the review if needed. Just let me know what version you're reviewing.

Also, we don't insist on using semantic versioning.
