+++ paper/paper.md
@@ -12,11 +12,11 @@ authors:
     orcid: 0000-0003-4552-5997
     affiliation: "1, 2, 3"
 affiliations:
- - name: Department of Physics and Astronomy, Macquarie University, Sydney, NSW 2109, Australia
+ - name: Research Centre in Astronomy, Astrophysics & Astrophotonics, Macquarie University, Sydney, NSW 2109, Australia
    index: 1
  - name: Harvard-Smithsonian Center for Astrophysics, 60 Garden Street, Cambridge, MA 02138, USA 
    index: 2
- - name: Department of Astronomy, University of Michigan, Ann Arbor, MI 48109, USA 
+ - name: Department of Astronomy, University of Michigan, 1085 S. University Avenue, Ann Arbor, MI 48109, USA 
    index: 3
 date: 30 June 2018
 bibliography: paper.bib
@@ -31,7 +31,7 @@ and ionic abundances can be determined from _collisionally excited lines_ (CEL) 
 equilibrium equations using collision strengths ($\Omega_{ij}$) and transition probabilities ($A_{ij}$) of ions. 
 Moreover, physical conditions and chemical abundances can be calculated from 
 _recombination lines_ (RL) using effective recombination coefficients ($\alpha_{\rm eff}$) of ions. 
-The atomic data (e.g. $\Omega_{ij}$, $A_{ij}$, and $\alpha_{\rm eff}$) 
+The atomic data, i.e. $\Omega_{ij}$, $A_{ij}$, and $\alpha_{\rm eff}$, 
 are used to calculate line emissivities in nebular spectral analysis tools 
 (e.g. @Howarth:1981; @Shaw:1994; @Shaw:1998; @Luridiana:2015; @Howarth:2016; @Danehkar:2018b), 
 and photoionization codes (e.g. @Ferland:1998; @Kallman:2001; @Ercolano:2003; @Ercolano:2005; @Ercolano:2008). 
@@ -41,11 +41,11 @@ nebulae from collisionally excited and recombination
 lines (see e.g. @Danehkar:2013; @Danehkar:2014; @Danehkar:2014b; @Danehkar:2016; @Danehkar:2018a).
 
 ``AtomNeb`` is a database containing atomic data stored in the Flexible Image 
-Transport System (FITS; @Wells:1981; @Hanisch:2001; @Pence:2010) file format 
+Transport System (FITS) file format (@Wells:1981; @Hanisch:2001; @Pence:2010) 
 produced for nebular spectral analysis. 
 FITS tables provide easy access to atomic data for spectral analysis tools. Especially, 
 ``AtomNeb`` includes the atomic data for both the _collisional excitation_ and 
-_recombination_ process of ions typically observed in nebular astrophysics. 
+_recombination_ process of ions usually observed in nebular astrophysics. 
 The ``AtomNeb`` interface library is equipped with several application programming 
 interface (API) functions written in the Interactive Data Language (IDL) 
 for reading the atomic data from the ``AtomNeb`` FITS files. Furthermore, the ``AtomNeb`` 
@@ -56,8 +56,7 @@ an open-source free compiler for IDL codes.
 language were designed to easily read _collision strengths_ ($\Omega_{ij}$) 
 and _transition probabilities_ ($A_{ij}$) of given ions, which can be used to derive electron 
 temperatures, electron densities, and ionic abundances from measured fluxes of 
-collisionally excited lines in spectral analysis tools, and to simulate line emissivities 
-in photoionization codes. The CEL data include energy levels ($E_{j}$), 
+collisionally excited lines. The CEL data include energy levels ($E_{j}$), 
[…]
