### 2019-03-04 #898 @danehkar
Considering Semantic Versioning (**MAJOR.MINOR.PATCH**), the code did not have any MAJOR change such as incompatible API changes, and any MINOR changes (e.g. adding a new functionality) since it was submitted. It has some changes in comments for documentation, so I changed the PATCH version from v0.0.1 to v0.0.2. I added an MGUnit tester, while the program functions were not changed, so there are no major or minor changes.

@arfon Please update Version from v0.0.1 to v0.0.2.
