+++ paper/paper.bib
@@ -346,7 +346,7 @@ archivePrefix = "arXiv",
       eid = {899},
     pages = {899},
       doi = {10.21105/joss.00899},
-   adsurl = {http://adsabs.harvard.edu/abs/2018arXiv181201605D},
+   adsurl = {http://adsabs.harvard.edu/abs/2018JOSS....3..899D},
   adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
