### 2019-03-06 #898 @danehkar
@arfon I did the proof correction in the paper. You can check it here: [799](https://github.com/openjournals/joss-reviews/issues/799). Please regenerated the final pdf for the paper. I also made an archive of the reviewed software in Zenodo here: https://doi.org/10.5281/zenodo.2584420
