### 2019-03-06 #898 @arfon
@danehkar - could you check those potentially missing DOIs and see if they are the correct ones for some of your references?
