### 2019-03-06 #898 @danehkar
@arfon Those references with missing DOI are from **Astronomy & Astrophysics** and **Astronomy and Astrophysics Supplement**:
[1994A&A...282..999S](http://adsabs.harvard.edu/abs/1994A%26A...282..999S)
[1991A&A...251..680P](http://adsabs.harvard.edu/abs/1991A%26A...251..680P)
[1981A&AS...44..363W](http://adsabs.harvard.edu/abs/1981A%26AS...44..363W)
These two journals (**Astronomy & Astrophysics** and **Astronomy and Astrophysics Supplement**) did not assign any Digital Object Identifier (DOI) to their publications prior to 2001, and they still have not assigned any DOI to their papers published before 2001.

Those DOI identified by the robot **whedon** are NOT related to these 3 references:
- https://doi.org/10.1051/0004-6361:20020263 may be missing for title: Recombination coefficients for O II lines at nebular temperatures and densities
- https://doi.org/10.1016/0032-0633(92)90127-a may be missing for title: Total and effective radiative recombination coefficients
- https://doi.org/10.1117/12.959819 may be missing for title: FITS - a Flexible Image Transport System

I confirm that the references in `paper.bib` are correct, as they are BibTex entries from [adsabs.harvard.edu](http://adsabs.harvard.edu/).
